+++ marbles/app/templates/footer.html
@@ -5,19 +5,19 @@
 
             <div class="col-md-3">
                 <h6 class="text-uppercase font-weight-bold">
-                    <a class="navitem nav-link" href="/info">Site Info</a>
+                    <a class="navitem nav-link nav-custom" href="/info">Site Info</a>
                 </h6>
             </div>
 
             <div class="col-md-3">
                 <h6 class="text-uppercase font-weight-bold">
-                    <a class="navitem nav-link" href="/admin">Admin</a>
+                    <a class="navitem nav-link nav-custom" href="/admin">Admin</a>
                 </h6>
             </div>
 
             <div class="col-md-3">
                 <h6 class="text-uppercase font-weight-bold">
-                    <a class="navitem nav-link" href="/about">About</a>
+                    <a class="navitem nav-link nav-custom" href="/about">About</a>
                 </h6>
             </div>
