+++ marbles/app/templates/form-updateDataModal.html
@@ -57,7 +57,9 @@
                             {% endif %}
                         </div>
                     </div>
-                    {{ form.submit(class_='btn btn-primary') }}
+                    <div class="modal-footer">
+                        {{ form.submit(class_='btn btn-primary') }}
+                    </div>
                 </form>
             </div>
         </div>
