+++ marbles/app/templates/horizontalBarChart.html
@@ -1,7 +1,10 @@
 <script>
     names = {{ names|tojson|safe }};
     wins = {{ wins|tojson|safe }};
+
     Chart.defaults.global.defaultFontSize = 15;
+    Chart.defaults.global.title.fontSize = 18;
+    Chart.defaults.global.title.fontFamily = "'Baloo Paaji 2'";
 
     new Chart(document.getElementById("bar-chart-horizontal"), {
         type: 'horizontalBar',
@@ -14,20 +17,20 @@
                 "rgba(25, 25, 25, 0.4)",        // BJ
                 "rgba(60, 50, 156, 0.4)",       // BG
                 "rgba(5, 99, 10, 0.4)",         // GG
-                "rgba(175, 175, 175, 0.4)",     // WL
+                "rgba(150, 150, 150, 0.4)",     // WL
             ],
             borderColor: [
                 "rgba(25, 25, 25, 1)",          // BJ
                 "rgba(60, 50, 156, 1)",         // BG
                 "rgba(5, 99, 10, 1)",           // GG
-                "rgba(175, 175, 175, 1)",       // WL
+                "rgba(150, 150, 150, 1)",       // WL
             ],
-            borderWidth: [1, 1, 1, 1],
+            borderWidth: [1.5, 1.5, 1.5, 1.5],
             hoverBackgroundColor: [
                 "rgba(25, 25, 25, 0.7)",        // BJ
                 "rgba(60, 50, 156, 0.7)",       // BG
                 "rgba(5, 99, 10, 0.7)",         // GG
-                "rgba(175, 175, 175, 0.7)",     // WL
+                "rgba(150, 150, 150, 0.7)",     // WL
             ],
             label: 'Wins',
             data: wins
@@ -45,7 +48,13 @@
                     ticks: {
                         beginAtZero: true,
                         stepSize: 1,
-                        suggestedMax: 5
+                        suggestedMax: 5,
+                        fontSize: 18
+                    }
+                }],
+                yAxes: [{
+                    ticks: {
+                        fontSize: 18
                     }
                 }]
             }
