+++ marbles/app/templates/index.html
@@ -5,7 +5,7 @@
     <section class="fwh standings">
         <div class="container fwh-content">
 
-            <h1>Race Standings</h1>
+            <h1 class="section-title">Race Standings</h1>
 
             <canvas id="bar-chart-horizontal" width="1000" height="500"></canvas>
