+++ marbles/app/templates/info.html
@@ -2,10 +2,30 @@
 
 {% block content %}
 
-    <section class="fwh">
-        <div class="container">
-            <h1>Info</h1>
-        </div>
+    <section class="container fwh-text">
+        <h1 class='section-title'>Site Info</h1>
+
+        <p class='section-text'>Lorem ipsum dolor, sit amet consectetur adipisicing elit.
+            Consectetur deleniti iste omnis quae id libero repellat, doloribus rerum maiores voluptate
+            sequi possimus sapiente quasi, deserunt veniam, odio expedita nihil iure.</p>
+
+        <p class='section-text'>Lorem ipsum dolor, sit amet consectetur adipisicing elit.
+            Consectetur deleniti iste omnis quae id libero repellat, doloribus rerum maiores voluptate
+            sequi possimus sapiente quasi, deserunt veniam, odio expedita nihil iure.</p>
+
+        <p class='section-text'>Lorem ipsum dolor, sit amet consectetur adipisicing elit.
+            Consectetur deleniti iste omnis quae id libero repellat, doloribus rerum maiores voluptate
+            sequi possimus sapiente quasi, deserunt veniam, odio expedita nihil iure.</p>
+
+        <hr>
+
+        <h1 class="section-title">
+            More Information
+        </h1>
+
+        <p class='section-text'>Lorem ipsum dolor, sit amet consectetur adipisicing elit.
+            Consectetur deleniti iste omnis quae id libero repellat, doloribus rerum maiores voluptate
+            sequi possimus sapiente quasi, deserunt veniam, odio expedita nihil iure.</p>
     </section>
 
 {% endblock %}
