+++ marbles/app/templates/navbar.html
@@ -1,4 +1,4 @@
-<nav class="navbar navbar-expand-lg navbar-light bg-light">
+<nav class="navbar navbar-expand-lg bg-dark">
     <a href="/" class="navbar-brand"><img class="logo" src="../static/images/marbleracing_logo.png" alt="Logo"></a>
     <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbar" aria-controls="navbarSupported">
         <span class="navbar-toggler-icon"></span>
@@ -8,17 +8,17 @@
         <div class="navbar-nav">
 
             {% if current_user.is_authenticated %}
-                <a href="/admin" class="navitem nav-link">Admin</a>
+                <a href="/admin" class="navitem nav-link nav-custom">Admin</a>
             {% endif %}
 
-            <a href="/" class="navitem nav-link">Home</a>
+            <a href="/" class="navitem nav-link nav-custom">Home</a>
 
-            <a href="/about" class="navitem nav-link">About</a>
+            <a href="/about" class="navitem nav-link nav-custom">About</a>
 
-            <a href="/data" class="navitem nav-link">Data</a>
+            <a href="/data" class="navitem nav-link nav-custom">Data</a>
 
             {% if current_user.is_authenticated %}
-                <a href="/logout" class="navitem nav-link">Logout</a>
+                <a href="/logout" class="navitem nav-link nav-custom">Logout</a>
             {% endif %}
 
         </div>
