+++ marbles/app/templates/styles.html
@@ -1,4 +1,11 @@
-<link rel="stylesheet" href="./static/bootstrap/dist/bootstrap.min.css">
+<!-- css -->
+<link rel="stylesheet" href="./static/bootstrap/css/bootstrap.min.css">
 <link rel="stylesheet" href="https://unpkg.com/bootstrap-table@1.16.0/dist/bootstrap-table.min.css">
 <link rel="stylesheet" href="../static/custom/css/custom.css">
+
+<!-- Icons -->
 <link rel="stylesheet" href="../static/fontawesome/css/all.min.css">
+
+
+<!-- Fonts -->
+<link href="https://fonts.googleapis.com/css2?family=Baloo+Paaji+2:wght@600&family=Poppins&display=swap" rel="stylesheet">
